requirement.txt
openai
streamlit
langchain
python-dotenv
PyPDF2
langchain-openai
reportlab